add fe,be,init module && add multiprocessing_pipe,redis_queue connector
