+++ src/cmd/remote-queue-chat/generate.py
@@ -1,0 +1,6 @@
+
+def main():
+    pass
+
+if __name__ == "__main__":
+    main()
